resources/python/destinations/websocket/requirements.txt
quixstreams>=3.0
python-dotenv
websockets>=14.0
asyncio